move dok to base
