+++ tests/DOK_test.cpp
@@ -1,4 +1,4 @@
-#include "solvant/sparse/DOK_matrix.hpp"
+#include "solvant/base/DOK_matrix.hpp"
 #include "gtest/gtest.h"
 
 using DOK_matrix  = solvant::sparse::DOK_matrix<double>;
